pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-resume-screener"
version = "0.1.0"
description = "FastAPI backend that parses resumes, stores structured data, and matches them to job descriptions using an LLM-driven scorer."
authors = [{name = "Your Name", email = "you@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "pdfminer.six>=20231228",
    "python-multipart>=0.0.9",
    "openai>=1.40.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0"
]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["."]
testpaths = ["tests"]
